requirements.txt
torch~=1.11.0
transformers~=4.18.0
attrs~=21.2.0
bs4~=0.0.1
beautifulsoup4~=4.10.0
chardet~=3.0.4
requests~=2.22.0
pandas~=1.3.3
sklearn~=0.0
scikit-learn~=0.24.2
numpy~=1.21.2
elasticsearch~=8.1.2
urllib3<1.26,>=1.21.1
nltk~=3.7
pyspellchecker~=0.6
tqdm~=4.64.0
hunspell~=0.5.5